Clean up and simplify MkDef() and MkNum().
Change MkDef() to just use a static buffer on the stack rather than
trying to dynamically allocate it.  4k is nearly 2 orders of magnitude
longer than any line can plausibly be, and is a trivial amount of
memory by 15 years ago standards.  If it IS exceeded, put a somewhat
meaningful debug line in the output.

Slightly increase the buffer in MkNum() (everybody knows all arbitrary
numbers should be powers of 2) and leave a comment about why it
Shouldn't Possibly overflow.  But switch to snprintf() to be sure it
can't be dangerous if it does.