Fix up losing the occupation window by iconifying it.
If you try to f.occupy something while the Occupation window is
already up, it skips out and does nothing.  But iconifying it doesn't
close it, it just hides the window.  So if you iconify the Occupation
window, you're screwed; it's lost forever.

Possibly we should just prevent iconifying it completely, but that's
more invasive, so the simplest thing is just to say "if it's not
mapped, it's not really up".  This does have the side effect that you
can now do things like f.addtoworkspace a window and have it take
effect while the Occupy window is up, which may confuse things a
little, but that's probably the best overall option...