Leave a comment about why I didn't rewrite this into a for() loop.
We can't actually just initialize to 0 and then check >'s in a loop,
since there are possible situations where the final answer could be <0
(imagine a set of RArea's off the left side of the screen).  The
situation's even worse for the Min[XY]2() variants, since there's no
practical "large than any possible" value to start from either.  So,
this pointer-stepping is the simplest way.