Eliminate grabbing the server in this section of startup.
It's a little unclear in looking if any of these bits really strongly
need to be atomic.  However, it's irrelevant, because grabbing doesn't
nest; calling XGrabServer() multiple times does nothing, and the first
XUngrabServer() call takes things to "normal".  Shortly into this
grabbed section, we loop over the windows on the screen, and
SimulateMapRequest() them all.  That HandleMapRequest()'s each, which
AddWindow()'s them.  And AddWindow() does a grab/ungrab pair.  So the
XUngrabServer() here actually never did anything, because the server
was already ungrab'd on the first window we picked up.

An argument could be made that AddWindow() should be more conditional
to _not_ ungrab during the startup.  The counterargument to that is
that it _has_ been doing so, and no visible bad effects have been
noticed (or been though important enough to investigate) in the last
couple decades that it's been doing that, so we might as well accept
it.