Create build option to work on disabling VirtualScreens.
This is unlikely to be remotely able to encompass all the code that
exists to support this, but at least we can knock out the
config-ability and a little top-level stuff easy enough.  Unwinding
all the mess in the code that exists to support them is its own long
task.