make indent
astyle seems to have changed some things in ways that aren't super
clear, so several lines are now indented somewhat differently.  I
can't run down any config params that have changed to cause this, or
any that sneak it back toward where it was, so just eat the changes
for now.